Refactor input file check for consistency across archive, copy, move, unarchive, unzst, zst-lite, and zst scripts

+++ commands/common/move.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 
-[ "$#" -eq 0 ] && echo -e "No input files" && return
+if [ $# -eq 0 ]; then
+  echo -e "No input files"
+  exit 1
+fi
 
 time rsync -rxzvuahP --remove-source-files --stats "${@}" && find "${1}" -type d -empty -delete
